Add public methods to support boxed text and arrows.
