+++ common/changes/@itwin/core-markup/markup-api_2023-04-04-16-11.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@itwin/core-markup",
+      "comment": "",
+      "type": "none"
+    }
+  ],
+  "packageName": "@itwin/core-markup"
+}
